clean up
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@2452 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/Graphos/GRBezierPath.h
@@ -66,7 +66,7 @@ typedef struct
 - (GRBezierControlPoint *)firstPoint;
 - (GRBezierControlPoint *)lastPoint;
 
-- (int)indexOfPoint:(GRBezierControlPoint *)aPoint;
+- (NSUInteger)indexOfPoint:(GRBezierControlPoint *)aPoint;
 
 
 @end
